Address PR #184 review feedback (CI fixes + cleanups)
Seven review items addressed in one commit (all pre-commit checks
re-run together → 771 unit + 18 integration tests green, zero
warnings, fmt clean):

- **AArch64 stochastic NZCV regression** (Codex bot inline +
  claude-bot critical #1, src/search/stochastic/backend.rs):
  `StochasticBackend<AArch64>::check_equivalence` now adds
  `.with_flags(true)` to the `EquivalenceConfig` it builds. Without
  this, the SMT solver could certify rewrites that change NZCV (e.g.
  `ADD;CMP` → `ADDS`) — the pre-PR `mcmc.rs` body had the flag
  explicitly and `SymbolicBackend<AArch64>` kept it. Phase C
  inadvertently dropped it during the genericisation; the commit
  message even predicted this would surface. Now fixed.

- **CI failure: unused `InstructionType` import** (clippy inline
  reports + `test` + `coverage` job failures; src/isa/x86.rs:1483 and
  :1534): the test modules import `ISAMutator, InstructionType` but
  use `instr.destination()` / `instr.source_registers()` which are
  inherent on `X86Instruction`. Trim the import to `ISAMutator` only.
  This unblocks CI (build is `-D warnings`).

- **clippy `doc_lazy_continuation`** (inline,
  src/parser/x86.rs:8): a wrapped doc line beginning with `+ #77`
  was read as a markdown list item. Rephrased to drop the leading
  `+`.

- **`SearchResultFor<I>` lived in the wrong module** (claude-bot
  major #6): defined in `stochastic/mcmc.rs` but consumed by
  `symbolic/synthesis.rs`. Moved to `search/result.rs` alongside
  `SearchResult`, with the `From<SearchResultFor<AArch64>>` impl.
  Both modules now import from the canonical home.

- **`make_cmp` dead duplicate** (claude-bot minor #9,
  src/parser/x86.rs): the `make_cmp` closure was character-for-
  character identical to `make`. Deleted; the `cmp` arm now uses
  `make` directly.

- **`default_x86_registers()` duplication** (claude-bot minor #10):
  `SearchConfig::default()` inlined the same 8-register pool that
  `search::candidate_x86::default_x86_registers()` already returns.
  Now calls the helper.

- **Compile-time generics test missed x86** (claude-bot nitpick #12,
  src/search/mod.rs): the `search_algorithm_is_generic_over_isa`
  smoke test only asserted AArch64 impls. Added four assertions for
  `StochasticSearch<X86_64>`, `SymbolicSearch<X86_64>`,
  `StochasticSearch<X86_32>`, `SymbolicSearch<X86_32>` so a future
  dropped backend impl fails to compile.

Remaining review items (`mutate_opcode` CMP no-op, hardcoded
`smt_timeout`, x86-32 width default trap, verbose-mode instr text,
`mutate_swap` rejection sampling, target-derived register pools,
stronger tracer-bullet assertions) are filed as separate tracking
issues — see the PR conversation reply for the full breakdown.
