+++ src/search/stochastic/mcmc.rs
@@ -16,7 +16,7 @@
 use crate::ir::Instruction;
 use crate::isa::{ISA, ISAMutator};
 use crate::search::config::SearchConfig;
-use crate::search::result::{SearchResult, SearchStatistics};
+use crate::search::result::{SearchResultFor, SearchStatistics};
 use crate::search::stochastic::acceptance::AcceptanceCriterion;
 use crate::search::stochastic::backend::StochasticBackend;
 use crate::search::{Algorithm, SearchAlgorithm};
@@ -267,67 +267,6 @@ where
 
     fn reset(&mut self) {
         self.statistics = SearchStatistics::new(Algorithm::Stochastic);
-    }
-}
-
-/// Generic search-result type. For AArch64, callers can ignore the
-/// type parameter (it defaults to `AArch64`) and treat
-/// `SearchResultFor<AArch64>` as the historical `SearchResult`.
-///
-/// Mirrors `crate::search::result::SearchResult` for `<I>`.
-#[derive(Debug, Clone)]
-pub struct SearchResultFor<I: ISA> {
-    pub optimized_sequence: Option<Vec<I::Instruction>>,
-    pub original_sequence: Vec<I::Instruction>,
-    pub found_optimization: bool,
-    pub statistics: SearchStatistics,
-}
-
-impl<I: ISA> SearchResultFor<I> {
-    /// Cost savings = original length minus optimized length, or 0 if
-    /// no optimization was found. Mirrors `SearchResult::cost_savings`.
-    pub fn cost_savings(&self) -> i64 {
-        if let Some(ref opt) = self.optimized_sequence {
-            self.original_sequence.len() as i64 - opt.len() as i64
-        } else {
-            0
-        }
-    }
-
-    pub fn no_optimization(original: Vec<I::Instruction>, statistics: SearchStatistics) -> Self {
-        Self {
-            optimized_sequence: None,
-            original_sequence: original,
-            found_optimization: false,
-            statistics,
-        }
-    }
-
-    pub fn with_optimization(
-        original: Vec<I::Instruction>,
-        optimized: Vec<I::Instruction>,
-        statistics: SearchStatistics,
-    ) -> Self {
-        Self {
-            optimized_sequence: Some(optimized),
-            original_sequence: original,
-            found_optimization: true,
-            statistics,
-        }
-    }
-}
-
-/// Backward-compatible conversion from the generic result type into the
-/// AArch64-specific `SearchResult`. Used by the parallel coordinator
-/// (which is still AArch64-typed) and any consumer that hasn't been
-/// migrated to the generic shape.
-impl From<SearchResultFor<crate::isa::AArch64>> for SearchResult {
-    fn from(r: SearchResultFor<crate::isa::AArch64>) -> Self {
-        if let Some(opt) = r.optimized_sequence {
-            SearchResult::with_optimization(r.original_sequence, opt, r.statistics)
-        } else {
-            SearchResult::no_optimization(r.original_sequence, r.statistics)
-        }
     }
 }
